Move stripManifestMetadata to diff.ts
