+++ ui/user/src/routes/admin/mcp-servers/c/[id]/instance/[ms_id]/+page.svelte
@@ -7,6 +7,7 @@
 	import McpServerActions from '$lib/components/mcp/McpServerActions.svelte';
 	import { VirtualPageViewport } from '$lib/components/ui/virtual-page';
 	import { DEFAULT_MCP_CATALOG_ID, PAGE_TRANSITION_DURATION } from '$lib/constants';
+	import { stripManifestMetadata } from '$lib/diff';
 	import { parseErrorContent } from '$lib/errors';
 	import {
 		type MCPCatalogEntryServerManifest,
@@ -14,7 +15,6 @@
 		type MCPCatalogServer,
 		AdminService
 	} from '$lib/services';
-	import { stripManifestMetadata } from '$lib/services/user/mcp';
 	import { mcpServersAndEntries, profile } from '$lib/stores';
 	import { CircleFadingArrowUp, Info, GitCompare } from 'lucide-svelte';
 	import { type Component, untrack } from 'svelte';
